requirements.txt
flask>=2.0.0
requests>=2.25.0
pyyaml>=5.4.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
